requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastapi==0.115.8
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.23
langchain-community==0.3.17
langchain-core==0.3.63
langchain-google-genai==2.1.2
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-checkpoint-redis==0.0.5
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langgraph-supervisor==0.0.27
langsmith==0.3.43
marshmallow==3.26.1
matplotlib-inline==0.1.7
ml_dtypes==0.5.1
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
ollama==0.5.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.8
ply==3.11
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
PyMuPDF==1.26.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ulid==3.0.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.3.0
redisvl==0.5.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
tabulate==0.9.0
tenacity==9.1.2
tornado==6.5.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
